[admin] Restrict recoverlist_view to superusers
django-reversion has a view where staff users
can recover deleted objects. The problem is that
it is not possible to filter these deleted objects
by organizations, therefore to avoid operators will
see a mix of their objects and objects of other organizations
we are restricting this view to superusers only.