[config/models] Fixed Error saving Device without organization #24
In the config class, device.organization was assigned to organization
without checking if the device actually had the organization or not.
Performing this check before assignment fixed the issue.
To test this change, an automated test was added which entails the
creation of a config_device with a device that has no organization.
a ValidationError is expected else the test fails.

Fixes #24