[geo/admin] Fixed 500 error when adding a new device #50
This 500 error was caused by a KeyError exception that was not catched
in the clean and cleaned_floorplan methods of AbstractObjectLocationForm
found in django_loci/base/admin.py. The problem was solved by overriding
these methods in geo/admin/ObjectLocationForm and thus catching the
exception in order to handle it by raising a ValidationError exception
to the form.
To test this changes, an automated test was created which replicates the
client post senario. If a KeyError is raise, the test is passed else the
test fails.

Fixes #50