[change] Avoid need of DJANGO_X509_CA_MODEL, DJANGO_X509_CERT_MODEL
Let's keep things simple for our users.
The pki models are the default ones, there's no need to require them
to set the setting explicitly.

This also simplifies things in ansible-openwisp2, docker-openwisp,
openwisp-monitoring and openwisp-firmware-upgrader, since we won't
have to update the code there, things will just keep working normally.

I think this is the approach to follow whenever a new setting is added,
we should provide a default value and avoid having he user to define it
explicitly.