[fix] Separated tests bound to development environment
Removed settings specifically set for running tests in settings.py
and mocked then on per test basis.

After upgrading openwisp-controller in anisble-openwisp2 role, the build was failing due to failing test cases. After debugging, I found out that there are specific settings which are set in test project of openwisp-controller which are absent from settings.py template in ansible-openwisp2.

The separation performed in this change allows to fix that problem cleanly.