[fix] Update related config when template default_values are changed
The config related to a template were not being flagged for update
when default_values were changed.
Moreover, during debugging, I found out that in some cases
the action was being performed even if there were no real changes,
the only difference being the new value was an instance of dict
while the value loaded from the model is an instance of OrderedDict.

This change fixes both issues.