[fix] Fixed unsaved_changes.js when location is present #388
Location geometry was not being parsed from JSON and hence
the geometry JSON representation had some slight differences
which where triggering the unsaved changes alert.

I took advantage to simplify the code and to fix the
non camelcase functions and variables in this part of the code.

I also updated a global variable name strating with _njc
(netjsonconfig) to owc (openwisp-controller).

Fixes #388