[change] Made updating Device reversion follow updatable #577
DeviceAdmin is used by other modules that register InlineModelAdmin
using monkey patching. The default implementation of reversion.register
ignore such inlines and does not update the "follow" field accordingly.

Thus, a classmethod is added in  DeviceAdmin to update the
"follow" fields of the Device model. This is achieved by unregistering
the Device model from reversion and re-registering it with updated
"follow" fields without altering other options.

Related to #577