[fix] Fix object and config menu not opening in Device config editor #910
This addresses the problem, that the drop-down for Configuration Menu
and Object Properties are not shown anymore in the device config
editor, if a configuration section is saved, which has in its JSON
schema definition a property with type array and should have unique
items. The cause of the problem is a call to super, when there is no
super. This commit will fix this issue by removing the call to super,
because it is not needed in this special case neither in all other
cases.

Fixes #910

Signed-off-by: Konrad Kreitmair <kkreitmair@tdt.de>