[fix] Fixed sorting issue in templates
Using the 'sort_value' of the 'through_model', created a subquery
which annotates a sort_value and selected field to each template.
The sort_value is used for ordering the templates and also for
checking if template is selected or not using a large default value
which  happens if template is not selected for a device config or
group.

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>