[fix] Fixed unsaved changes alert for readonly map #560
For readonly mode the field name used to track geometry field
is different leading to its json value not being parsed, hence
leading to unsaved_changes alert. A small change of addition
of the said field in 'jsonValues' of 'unsaved_changes.js' Fixed
it.

Also increased the timeout for waiting of success message present
for selenium test of device admin to tackle TimeoutException.

Closes #560

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>
Co-authored-by: Federico Capoano <f.capoano@openwisp.io>