[tests] Added tests for WHOIS creation #1045
Added tests for checking if `fetch_whois_details` task
is called properly or not, and for checking creation
of WHOIS record for a device if the last_ip recorded
is public.
As now whois record is also fetched whenever fetching
a device, there is an increase in query count in some
of the tests.

Closes #1045

Signed-off-by: DragnEmperor <dragnemperor@gmail.com>